requirements.txt
alembic==1.4.3
amqp==2.6.1
appdirs==1.4.4
appnope==0.1.0
atomicwrites==1.4.0
attrs==19.3.0
awsebcli==3.19.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
billiard==3.5.0.5
botocore==1.17.62
captcha==0.3
celery==4.2.1
cement==2.8.2
certifi==2019.3.9
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2.1
cycler==0.10.0
decorator==4.1.2
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2
django-cors-headers==3.5.0
django-heroku==0.3.1
django-paypal==1.0.0
django-rest-swagger==2.2.0
django-restframework==0.0.1
djangorestframework==3.9.2
docutils==0.15.2
enum34==1.1.6
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.3
Flask-MongoAlchemy==0.7.2
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.4
future==0.16.0
gunicorn==19.9.0
html-testRunner==1.2.1
idna==2.7
iniconfig==1.1.1
ipdb==0.10.3
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kiwisolver==1.2.0
kombu==4.6.11
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mock==4.0.2
MongoAlchemy==0.19
more-itertools==8.4.0
mysql==0.0.2
mysqlclient==2.0.1
nose==1.3.7
numpy==1.19.2
openapi-codec==1.3.2
packaging==20.4
pathlib2==2.3.0
pathspec==0.5.9
paypal==1.2.5
pexpect==4.3.0
pickleshare==0.7.4
Pillow==7.2.0
pip-conflict-checker==0.6.0
pipdeptree==1.0.0
pipenv==2018.11.26
pluggy==0.13.1
prompt-toolkit==1.0.15
psycopg2==2.8.6
ptyprocess==0.5.2
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.8
pyflakes==2.2.0
Pygments==2.2.0
pymongo==3.11.0
PyMySQL==0.10.1
pyparsing==2.4.7
pypiwin32==223
pytest==6.1.2
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
pywin32==228
PyYAML==5.3.1
rabbitmq==0.2.0
requests==2.20.1
selenium==3.141.0
semantic-version==2.5.0
simplegeneric==0.8.1
simplejson==3.17.2
six==1.11.0
SQLAlchemy==1.3.18
sqlparse==0.3.0
termcolor==1.1.0
ticket-book-pkg-jejurkarganesh1990 @ file:///F:/work/P10/gbj_ticket_app/dist/ticket_book_pkg_jejurkarganesh1990-5.5.1-py3-none-any.whl
toml==0.10.2
traitlets==4.3.2
uritemplate==3.0.1
urllib3==1.24.3
vine==1.3.0
virtualenv==16.5.0
virtualenv-clone==0.5.3
wcwidth==0.1.7
Werkzeug==1.0.1
whitenoise==4.1.2
